+++ canvas_sdk/v1/data/__init__.py
@@ -42,6 +42,12 @@ from .claim import (
 from .claim_banner_alert import BannerAlertIntent, BannerAlertStatus, ClaimBannerAlert
 from .claim_diagnosis_code import ClaimDiagnosisCode
 from .claim_line_item import ClaimLineItem, ClaimLineItemDiagnosisCode, ClaimLineItemModifier
+from .coding_gap_event import (
+    AssessCodingGapEvent,
+    CreateCodingGapEvent,
+    DeferCodingGapEvent,
+    ValidateCodingGapEvent,
+)
 from .command import Command, CommandMetadata
 from .compound_medication import CompoundMedication
 from .condition import Condition, ConditionCoding
@@ -77,6 +83,7 @@ from .facility import Facility
 from .family_history import FamilyHistory, FamilyHistoryCoding
 from .follow_up import FollowUp
 from .goal import Goal, UpdateGoal
+from .group import Group
 from .history_present_illness import HistoryOfPresentIllness
 from .imaging import (
     ImagingOrder,
@@ -168,6 +175,7 @@ from .patient_consent import (
     PatientConsentRejectionCoding,
 )
 from .patient_group import PatientGroup, PatientGroupMember
+from .payment_card import PaymentCard
 from .payment_collection import PaymentCollection
 from .payor_specific_charge import PayorSpecificCharge
 from .plan import Plan
@@ -203,12 +211,14 @@ from .questionnaire import (
     InterviewQuestionnaireMap,
     InterviewQuestionResponse,
     Question,
+    QuestionEnablementCondition,
     Questionnaire,
     QuestionnaireQuestionMap,
     ResponseOption,
     ResponseOptionSet,
 )
 from .reason_for_visit import ReasonForVisit, ReasonForVisitCoding, ReasonForVisitSettingCoding
+from .reference import Reference
 from .referral import Referral, ReferralReport, ReferralReportCoding, ReferralReview
 from .refill_request import RefillRequest, RefillRequestCoding
 from .remove_allergy_event import RemoveAllergyEvent
@@ -250,6 +260,7 @@ __all__ = __exports__ = (
     "AppointmentLabel",
     "AllergyIntolerance",
     "AllergyIntoleranceCoding",
+    "AssessCodingGapEvent",
     "Assessment",
     "AttributeHub",
     "BannerAlert",
@@ -299,7 +310,9 @@ __all__ = __exports__ = (
     "ModelExtension",
     "CustomModel",
     "CoveragePosting",
+    "CreateCodingGapEvent",
     "CurrentNoteStateEvent",
+    "DeferCodingGapEvent",
     "DetectedIssue",
     "DetectedIssueEvidence",
     "Device",
@@ -330,6 +343,7 @@ __all__ = __exports__ = (
     "FieldType",
     "FollowUp",
     "Goal",
+    "Group",
     "HistoryOfPresentIllness",
     "ImagingOrder",
     "ImagingReport",
@@ -424,6 +438,7 @@ __all__ = __exports__ = (
     "PatientGroup",
     "PatientGroupMember",
     "PayorSpecificCharge",
+    "PaymentCard",
     "PaymentCollection",
     "Plan",
     "PluginCommand",
@@ -447,11 +462,13 @@ __all__ = __exports__ = (
     "ProtocolCurrent",
     "ProtocolOverride",
     "Question",
+    "QuestionEnablementCondition",
     "Questionnaire",
     "QuestionnaireQuestionMap",
     "ReasonForVisit",
     "ReasonForVisitCoding",
     "ReasonForVisitSettingCoding",
+    "Reference",
     "Referral",
     "ReferralReport",
     "ReferralReportCoding",
@@ -493,6 +510,7 @@ __all__ = __exports__ = (
     "Vaccine",
     "VaccineLot",
     "VaccineManufacturer",
+    "ValidateCodingGapEvent",
     "VisualExamFinding",
     "VitalSign",
     "VitalSignReading",